Create Multipleinifiniboxpoolcapacityreporting.py
This script connects to the infinidat infinibox storage using readonly credentials and makes api get calls for the pool information which is then stored in a mysql database(in this case the mysqldb is hosted on the same server as the script runs). You will notice 3 URLs, one for each storage array. If you have a new storage array all you need to do is create another variable lets say of this script it will be urld, since urlc is the previous one. Then you can add the api link for the pools into it as a string. Do not forget to add the urld into the list urllist variable mentioned in line 22(in this script). 
This script automatically pulls the storage array name from your url and adds into the database column. This script also collects the current date as of when you run the script and also enters into the database.                                                  
You are free to copy this script and modify it or enhance it. Please share your enhancements so it can help others and myself learn